triples_batched -> triples
