+++ mrmustard/lab_dev/states/displaced_squeezed.py
@@ -19,7 +19,7 @@ The class representing a displaced squeezed state.
 from __future__ import annotations
 
 from mrmustard.physics.ansatz import PolyExpAnsatz
-from mrmustard.physics import triples_batched
+from mrmustard.physics import triples
 from .ket import Ket
 from ..utils import make_parameter
 
@@ -80,7 +80,7 @@ class DisplacedSqueezed(Ket):
         self._representation = self.from_ansatz(
             modes=(mode,),
             ansatz=PolyExpAnsatz.from_function(
-                fn=triples_batched.displaced_squeezed_vacuum_state_Abc,
+                fn=triples.displaced_squeezed_vacuum_state_Abc,
                 x=self.parameters.x,
                 y=self.parameters.y,
                 r=self.parameters.r,
