+++ mrmustard/lab_dev/transformations/s2gate.py
@@ -20,7 +20,7 @@ from __future__ import annotations
 
 from .base import Unitary
 from ...physics.ansatz import PolyExpAnsatz
-from ...physics import triples_batched
+from ...physics import triples
 from ..utils import make_parameter
 
 __all__ = ["S2gate"]
@@ -81,7 +81,7 @@ class S2gate(Unitary):
             modes_in=modes,
             modes_out=modes,
             ansatz=PolyExpAnsatz.from_function(
-                fn=triples_batched.twomode_squeezing_gate_Abc,
+                fn=triples.twomode_squeezing_gate_Abc,
                 r=self.parameters.r,
                 phi=self.parameters.phi,
             ),
