+++ mrmustard/physics/representations.py
@@ -27,7 +27,7 @@ from mrmustard.utils.typing import (
 )
 
 from .ansatz import Ansatz, PolyExpAnsatz, ArrayAnsatz
-from .triples_batched import identity_Abc
+from .triples import identity_Abc
 from .wires import Wires, ReprEnum
 
 __all__ = ["Representation"]
