+++ mrmustard/physics/triples.py
@@ -18,18 +18,41 @@ various states and transformations.
 """
 from __future__ import annotations
 
-from typing import Generator, Iterable, Union
+from typing import Iterable
 
 import numpy as np
 
 from mrmustard import math, settings
-from mrmustard.utils.typing import Matrix, Vector, Scalar, RealMatrix, ComplexMatrix
+from mrmustard.utils.typing import Matrix, Vector, Scalar, RealMatrix
 from mrmustard.physics.gaussian_integrals import complex_gaussian_integral_2
 from .bargmann_utils import symplectic2Au
+
 
 #  ~~~~~~~~~
 #  Utilities
 #  ~~~~~~~~~
+
+
+def _compute_batch_size(*args) -> tuple[int, ...] | None:
+    r"""
+    Compute the final batch size of the input arguments.
+
+    Args:
+        *args: The input arguments.
+
+    Returns:
+        The final batch size and batch dimension of the input arguments.
+    """
+    batch_size = None
+    for arg in args:
+        arg = math.astensor(arg)
+        if arg.shape:
+            if batch_size is None:
+                batch_size = arg
+            else:
+                batch_size = batch_size * arg
+    batch_size = batch_size.shape if batch_size is not None else None
+    return batch_size, len(batch_size or (1,))
 
 
 def _X_matrix_for_unitary(n_modes: int) -> Matrix:
@@ -53,32 +76,12 @@ def _vacuum_B_vector(n_modes: int) -> Vector:
     return math.zeros((n_modes,), math.complex128)
 
 
-def _reshape(**kwargs) -> Generator:
-    r"""
-    A utility function to reshape parameters.
-    """
-    names = list(kwargs.keys())
-    vars = list(kwargs.values())
-
-    vars = [math.atleast_1d(var, math.complex128) for var in vars]
-    n_modes = max(len(var) for var in vars)
-
-    for i, var in enumerate(vars):
-        if len(var) == 1:
-            var = math.tile(var, (n_modes,))
-        else:
-            if len(var) != n_modes:
-                msg = f"Parameter {names[i]} has an incompatible shape."
-                raise ValueError(msg)
-        yield var
-
-
 #  ~~~~~~~~~~~
 #  Pure States
 #  ~~~~~~~~~~~
 
 
-def vacuum_state_Abc(n_modes: int) -> Union[Matrix, Vector, Scalar]:
+def vacuum_state_Abc(n_modes: int) -> tuple[Matrix, Vector, Scalar]:
     r"""
     The ``(A, b, c)`` triple of a tensor product of vacuum states on ``n_modes``.
 
@@ -95,90 +98,85 @@ def vacuum_state_Abc(n_modes: int) -> Union[Matrix, Vector, Scalar]:
     return A, b, c
 
 
-def bargmann_eigenstate_Abc(x: Union[float, Iterable[float]]) -> Union[Matrix, Vector, Scalar]:
+def bargmann_eigenstate_Abc(alpha: complex | Iterable[complex]) -> tuple[Matrix, Vector, Scalar]:
     r"""
     The Abc triple of a Bargmann eigenstate.
+
+    Args:
+        alpha: The eigenvalue of the Bargmann eigenstate.
+
+    Returns:
+        The ``(A, b, c)`` triple of the Bargmann eigenstate.
     """
-    x = list(_reshape(x=x))
-    nmodes = len(x[0])
-    A = _vacuum_A_matrix(nmodes)
-    b = x
-    c = 1
-    return A, b, c
+    batch_size, _ = _compute_batch_size(alpha)
+    batch_shape = batch_size or (1,)
+
+    A = math.broadcast_to(_vacuum_A_matrix(1), batch_shape + (1, 1))
+    b = math.cast(math.reshape(alpha, batch_shape + (1,)), math.complex128)
+    c = math.ones(batch_shape, math.complex128)
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def coherent_state_Abc(
-    x: Union[float, Iterable[float]], y: Union[float, Iterable[float]] = 0
-) -> Union[Matrix, Vector, Scalar]:
+    x: float | Iterable[float], y: float | Iterable[float] = 0
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of pure coherent states.
-
-    The number of modes depends on the length of the input parameters.
-
-    If one of the input parameters has length ``1``, it is tiled so that its length matches
-    that of the other one. For example, passing ``x=[1,2,3]`` and ``y=1`` is equivalent to passing
-    ``x=[1,2,3]`` and ``y=[1,1,1]``.
+    The ``(A, b, c)`` triple of a pure coherent state.
 
     Args:
-        x: The real parts of the displacements, in units of :math:`\sqrt{\hbar}`.
-        y: The imaginary parts of the displacements, in units of :math:`\sqrt{\hbar}`.
+        x: The real part of the displacement, in units of :math:`\sqrt{\hbar}`.
+        y: The imaginary part of the displacement, in units of :math:`\sqrt{\hbar}`.
 
     Returns:
-        The ``(A, b, c)`` triple of the pure coherent states.
+        The ``(A, b, c)`` triple of the pure coherent state.
     """
-    x, y = list(_reshape(x=x, y=y))
-    n_modes = len(x)
+    batch_size, _ = _compute_batch_size(x, y)
+    batch_shape = batch_size or (1,)
 
-    A = _vacuum_A_matrix(n_modes)
-    b = x + 1j * y
-    c = math.prod(math.exp(-0.5 * (x**2 + y**2)))
+    x = math.broadcast_to(x, batch_shape, dtype=math.complex128)
+    y = math.broadcast_to(y, batch_shape, dtype=math.complex128)
 
-    return A, b, c
+    A = math.broadcast_to(_vacuum_A_matrix(1), batch_shape + (1, 1))
+    b = math.reshape(x + 1j * y, batch_shape + (1,))
+    c = math.cast(math.exp(-0.5 * (x**2 + y**2)), math.complex128)
+
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def squeezed_vacuum_state_Abc(
-    r: Union[float, Iterable[float]], phi: Union[float, Iterable[float]] = 0
-) -> Union[Matrix, Vector, Scalar]:
+    r: float | Iterable[float], phi: float | Iterable[float] = 0
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of squeezed vacuum states.
-
-    The number of modes depends on the length of the input parameters.
-
-    If one of the input parameters has length ``1``, it is tiled so that its length matches
-    that of the other one. For example, passing ``r=[1,2,3]`` and ``phi=1`` is equivalent to
-    passing ``r=[1,2,3]`` and ``phi=[1,1,1]``.
+    The ``(A, b, c)`` triple of a squeezed vacuum state.
 
     Args:
         r: The squeezing magnitudes.
         phi: The squeezing angles.
 
     Returns:
-        The ``(A, b, c)`` triple of the squeezed vacuum states.
+        The ``(A, b, c)`` triple of a squeezed vacuum state.
     """
-    r, phi = list(_reshape(r=r, phi=phi))
-    n_modes = len(r)
+    batch_size, _ = _compute_batch_size(r, phi)
+    batch_shape = batch_size or (1,)
 
-    A = math.diag(-math.sinh(r) / math.cosh(r) * math.exp(1j * phi))
-    b = _vacuum_B_vector(n_modes)
-    c = math.prod(1 / math.sqrt(math.cosh(r)))
+    r = math.broadcast_to(r, batch_shape, dtype=math.complex128)
+    phi = math.broadcast_to(phi, batch_shape, dtype=math.complex128)
 
-    return A, b, c
+    A = math.reshape(-math.sinh(r) / math.cosh(r) * math.exp(1j * phi), batch_shape + (1, 1))
+    b = math.broadcast_to(_vacuum_B_vector(1), batch_shape + (1,))
+    c = 1 / math.sqrt(math.cosh(r))
+
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def displaced_squeezed_vacuum_state_Abc(
-    x: Union[float, Iterable[float]],
-    y: Union[float, Iterable[float]] = 0,
-    r: Union[float, Iterable[float]] = 0,
-    phi: Union[float, Iterable[float]] = 0,
-) -> Union[Matrix, Vector, Scalar]:
+    x: float | Iterable[float],
+    y: float | Iterable[float] = 0,
+    r: float | Iterable[float] = 0,
+    phi: float | Iterable[float] = 0,
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of displazed squeezed vacuum states.
-
-    The number of modes depends on the length of the input parameters.
-
-    If some of the input parameters have length ``1``, they are tiled so that their length
-    matches that of the other ones. For example, passing ``r=[1,2,3]`` and ``phi=1`` is equivalent
-    to passing ``r=[1,2,3]`` and ``phi=[1,1,1]``.
+    The ``(A, b, c)`` triple of a displaced squeezed vacuum state.
 
     Args:
         r: The squeezing magnitudes.
@@ -187,50 +185,60 @@ def displaced_squeezed_vacuum_state_Abc(
         y: The imaginary parts of the displacements, in units of :math:`\sqrt{\hbar}`.
 
     Returns:
-        The ``(A, b, c)`` triple of the squeezed vacuum states.
+        The ``(A, b, c)`` triple of the squeezed vacuum state.
     """
-    x, y, r, phi = list(_reshape(x=x, y=y, r=r, phi=phi))
+    batch_size, _ = _compute_batch_size(x, y, r, phi)
+    batch_shape = batch_size or (1,)
 
-    A = math.diag(-math.sinh(r) / math.cosh(r) * math.exp(1j * phi))
-    b = (x + 1j * y) + (x - 1j * y) * math.sinh(r) / math.cosh(r) * math.exp(1j * phi)
+    x = math.broadcast_to(x, batch_shape, dtype=math.complex128)
+    y = math.broadcast_to(y, batch_shape, dtype=math.complex128)
+    r = math.broadcast_to(r, batch_shape, dtype=math.complex128)
+    phi = math.broadcast_to(phi, batch_shape, dtype=math.complex128)
+
+    A = math.reshape(-math.sinh(r) / math.cosh(r) * math.exp(1j * phi), batch_shape + (1, 1))
+    b = math.reshape(
+        (x + 1j * y) + (x - 1j * y) * math.sinh(r) / math.cosh(r) * math.exp(1j * phi),
+        batch_shape + (1,),
+    )
     c = math.exp(
         -0.5 * (x**2 + y**2)
         - 0.5 * (x - 1j * y) ** 2 * math.sinh(r) / math.cosh(r) * math.exp(1j * phi)
     )
-    c = math.prod(c / math.sqrt(math.cosh(r)))
+    c = c / math.sqrt(math.cosh(r))
 
-    return A, b, c
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def two_mode_squeezed_vacuum_state_Abc(
-    r: Union[float, Iterable[float]], phi: Union[float, Iterable[float]] = 0
-) -> Union[Matrix, Vector, Scalar]:
+    r: float | Iterable[float], phi: float | Iterable[float] = 0
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of two mode squeezed vacuum states.
-
-    The number of modes depends on the length of the input parameters.
-
-    If one of the input parameters has length ``1``, it is tiled so that its length matches
-    that of the other one. For example, passing ``r=[1,2,3,4]`` and ``phi=1`` is equivalent to
-    passing ``r=[1,2,3,4]`` and ``phi=[1,1,1,1]``.
+    The ``(A, b, c)`` triple of a two mode squeezed vacuum state.
 
     Args:
         r: The squeezing magnitudes.
         phi: The squeezing angles.
 
     Returns:
-        The ``(A, b, c)`` triple of the squeezed vacuum states.
+        The ``(A, b, c)`` triple of the squeezed vacuum state.
     """
-    r, phi = list(_reshape(r=r, phi=phi))
-    n_modes = 2 * len(r)
-    O = math.zeros((len(r), len(r)), math.complex128)
-    tanhr = math.diag(-math.exp(1j * phi) * math.sinh(r) / math.cosh(r))
+    batch_size, batch_dim = _compute_batch_size(r, phi)
+    batch_shape = batch_size or (1,)
 
-    A = math.block([[O, tanhr], [tanhr, O]])
-    b = _vacuum_B_vector(n_modes)
-    c = math.prod(1 / math.cosh(r))
+    r = math.broadcast_to(r, batch_shape, dtype=math.complex128)
+    phi = math.broadcast_to(phi, batch_shape, dtype=math.complex128)
 
-    return A, b, c
+    O_matrix = math.zeros(batch_shape, math.complex128)
+    tanhr = -math.exp(1j * phi) * math.sinh(r) / math.cosh(r)
+
+    A = math.stack(
+        [math.stack([O_matrix, tanhr], batch_dim), math.stack([tanhr, O_matrix], batch_dim)],
+        batch_dim,
+    )
+    b = math.broadcast_to(_vacuum_B_vector(2), batch_shape + (2,))
+    c = math.cast(1 / math.cosh(r), math.complex128)
+
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def gket_state_Abc(symplectic: RealMatrix):
@@ -243,15 +251,25 @@ def gket_state_Abc(symplectic: RealMatrix):
     Returns:
         The ``(A,b,c)`` triple of the Gket state.
     """
+    batch_size = symplectic.shape[:-2]
+    batch_shape = batch_size or (1,)
 
+    symplectic = math.broadcast_to(symplectic, batch_shape + symplectic.shape[-2:])
     m = symplectic.shape[-1] // 2  # num of modes
 
-    A = symplectic2Au(symplectic)
-    b = math.zeros(m, dtype=A.dtype)
-    c = ((-1) ** m * math.det(A[m:, m:] @ math.conj(A[m:, m:]) - math.eye_like(A[m:, m:]))) ** 0.25
-    return A[:m, :m], b, c
+    Au = symplectic2Au(symplectic)
+
+    A = Au[..., :m, :m]
+    b = math.zeros(batch_shape + (m,), dtype=A.dtype)
+    c = (
+        (-1) ** m
+        * math.det(Au[..., m:, m:] @ math.conj(Au[..., m:, m:]) - math.eye_like(Au[..., m:, m:]))
+    ) ** 0.25
+
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
+# TODO: update after complex_gaussian_integral_2 is updated
 def gdm_state_Abc(betas: Vector, symplectic: RealMatrix):
     r"""
     The A,b,c parameters of a Gaussian mixed state that is defined by the action of a Guassian on a thermal state
@@ -286,7 +304,8 @@ def gdm_state_Abc(betas: Vector, symplectic: RealMatrix):
     )
 
 
-def sauron_state_Abc(n: int, epsilon: float):
+# TODO: update after complex_gaussian_integral_2 is updated
+def sauron_state_Abc(n: int, epsilon: float) -> tuple[Matrix, Vector, Scalar]:
     r"""
     The A,b,c parametrization of Sauron states. These are Fock states written as a linear superposition of a
     ring of coherent states.
@@ -314,31 +333,31 @@ def sauron_state_Abc(n: int, epsilon: float):
     return As, bs, cs
 
 
-def quadrature_eigenstates_Abc(x: float, phi: float) -> Union[Matrix, Vector, Scalar]:
+def quadrature_eigenstates_Abc(
+    x: float | Iterable[float], phi: float | Iterable[float]
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of quadrature eigenstates.
-
-    The number of modes depends on the length of the input parameters.
-
-    If one of the input parameters has length ``1``, it is tiled so that its length matches
-    that of the other one. For example, passing ``x=[1,2,3]`` and ``phi=1`` is equivalent to
-    passing ``x=[1,2,3]`` and ``phi=[1,1,1]``.
+    The ``(A, b, c)`` triple of a quadrature eigenstate.
 
     Args:
         r: The squeezing magnitudes.
         phi: The squeezing angles.
 
     Returns:
-        The ``(A, b, c)`` triple of the squeezed vacuum states.
+        The ``(A, b, c)`` triple of the squeezed vacuum state.
     """
     hbar = settings.HBAR
-    x, phi = list(_reshape(x=x, phi=phi))
 
-    A = -math.diag(math.exp(1j * 2 * phi))
-    b = x * math.exp(1j * phi) * math.sqrt(2 / hbar)
-    c = math.prod(1 / (np.pi) ** (1 / 4) * math.exp(-(x**2) / (2 * hbar)))
+    batch_size, _ = _compute_batch_size(x, phi)
+    batch_shape = batch_size or (1,)
 
-    return A, b, c
+    x = math.broadcast_to(x, batch_shape, dtype=math.complex128)
+    phi = math.broadcast_to(phi, batch_shape, dtype=math.complex128)
+
+    A = math.reshape(-math.exp(1j * 2 * phi), batch_shape + (1, 1))
+    b = math.reshape(x * math.exp(1j * phi) * math.sqrt(2 / hbar), batch_shape + (1,))
+    c = math.cast(1 / (np.pi) ** (1 / 4) * math.exp(-(x**2) / (2 * hbar)), math.complex128)
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 #  ~~~~~~~~~~~~
@@ -346,27 +365,33 @@ def quadrature_eigenstates_Abc(x: float, phi: float) -> Union[Matrix, Vector, Sc
 #  ~~~~~~~~~~~~
 
 
-def thermal_state_Abc(nbar: Union[int, Iterable[int]]) -> Union[Matrix, Vector, Scalar]:
+def thermal_state_Abc(nbar: int | Iterable[int]) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of thermal states.
-
-    The number of modes depends on the length of the input parameters.
+    The ``(A, b, c)`` triple of a thermal state.
 
     Args:
-        nbar: The average numbers of photons per mode.
+        nbar: The average number of photons.
 
     Returns:
-        The ``(A, b, c)`` triple of the thermal states.
+        The ``(A, b, c)`` triple of the thermal state.
     """
-    nbar = math.atleast_1d(nbar, math.complex128)
-    n_modes = len(nbar)
+    batch_size, batch_dim = _compute_batch_size(nbar)
+    batch_shape = batch_size or (1,)
 
-    A = math.astensor([[0, 1], [1, 0]], math.complex128)
-    A = math.kron((nbar / (nbar + 1)) * math.eye(n_modes, math.complex128), A)
-    c = math.prod(math.astensor([1 / (_nbar + 1) for _nbar in nbar]))
-    b = _vacuum_B_vector(n_modes * 2)
+    nbar = math.broadcast_to(nbar, batch_shape, dtype=math.complex128)
 
-    return A, b, c
+    O_matrix = math.zeros(batch_shape, math.complex128)
+
+    A = math.stack(
+        [
+            math.stack([O_matrix, (nbar / (nbar + 1))], batch_dim),
+            math.stack([(nbar / (nbar + 1)), O_matrix], batch_dim),
+        ],
+        batch_dim,
+    )
+    b = math.broadcast_to(_vacuum_B_vector(2), batch_shape + (2,))
+    c = math.cast(1 / (nbar + 1), math.complex128)
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~
@@ -375,170 +400,189 @@ def thermal_state_Abc(nbar: Union[int, Iterable[int]]) -> Union[Matrix, Vector, 
 
 
 def rotation_gate_Abc(
-    theta: Union[float, Iterable[float]],
-) -> Union[Matrix, Vector, Scalar]:
+    theta: float | Iterable[float],
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of of a tensor product of rotation gates.
-
-    The number of modes depends on the length of the input parameters.
+    The ``(A, b, c)`` triple of of a tensor product of a rotation gate.
 
     Args:
         theta: The rotation angles.
 
     Returns:
-        The ``(A, b, c)`` triple of the rotation gates.
+        The ``(A, b, c)`` triple of the rotation gate.
     """
-    theta = math.atleast_1d(theta, math.complex128)
-    n_modes = len(theta)
+    batch_size, batch_dim = _compute_batch_size(theta)
+    batch_shape = batch_size or (1,)
 
-    A = math.astensor([[0, 1], [1, 0]], math.complex128)
-    A = math.kron(A, math.exp(1j * theta) * math.eye(n_modes, math.complex128))
-    b = _vacuum_B_vector(n_modes * 2)
-    c = 1.0 + 0j
+    theta = math.broadcast_to(theta, batch_shape, dtype=math.complex128)
 
-    return A, b, c
+    O_matrix = math.zeros(batch_shape, math.complex128)
+
+    A = math.stack(
+        [
+            math.stack([O_matrix, math.exp(1j * theta)], batch_dim),
+            math.stack([math.exp(1j * theta), O_matrix], batch_dim),
+        ],
+        batch_dim,
+    )
+    b = math.broadcast_to(_vacuum_B_vector(2), batch_shape + (2,))
+    c = math.ones(batch_shape, math.complex128)
+
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def displacement_gate_Abc(
-    x: Union[float, Iterable[float]], y: Union[float, Iterable[float]] = 0
-) -> Union[Matrix, Vector, Scalar]:
+    x: float | Iterable[float], y: float | Iterable[float] = 0
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of displacement gates.
-
-    The number of modes depends on the length of the input parameters.
-
-    If one of the input parameters has length ``1``, it is tiled so that its length matches
-    that of the other one. For example, passing ``x=[1,2,3]`` and ``y=1`` is equivalent to
-    passing ``x=[1,2,3]`` and ``y=[1,1,1]``.
+    The ``(A, b, c)`` triple of a tensor product of a displacement gate.
 
     Args:
-        x: The real parts of the displacements, in units of :math:`\sqrt{\hbar}`.
-        y: The imaginary parts of the displacements, in units of :math:`\sqrt{\hbar}`.
+        x: The real part of the displacement, in units of :math:`\sqrt{\hbar}`.
+        y: The imaginary part of the displacement, in units of :math:`\sqrt{\hbar}`.
 
     Returns:
-        The ``(A, b, c)`` triple of the displacement gates.
+        The ``(A, b, c)`` triple of the displacement gate.
     """
-    x, y = _reshape(x=x, y=y)
-    n_modes = len(x)
+    batch_size, batch_dim = _compute_batch_size(x, y)
+    batch_shape = batch_size or (1,)
 
-    A = _X_matrix_for_unitary(n_modes)
-    b = math.concat([x + 1j * y, -x + 1j * y], axis=0)
-    c = math.exp(-math.sum(x**2 + y**2) / 2)
+    x = math.broadcast_to(x, batch_shape, dtype=math.complex128)
+    y = math.broadcast_to(y, batch_shape, dtype=math.complex128)
 
-    return A, b, c
+    A = math.broadcast_to(_X_matrix_for_unitary(1), batch_shape + (2, 2))
+    b = math.stack([x + 1j * y, -x + 1j * y], batch_dim)
+    c = math.cast(math.exp(-(x**2 + y**2) / 2), math.complex128)
+
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def squeezing_gate_Abc(
-    r: Union[float, Iterable[float]], delta: Union[float, Iterable[float]] = 0
-) -> Union[Matrix, Vector, Scalar]:
+    r: float | Iterable[float], delta: float | Iterable[float] = 0
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of squeezing gates.
-
-    The number of modes depends on the length of the input parameters.
-
-    If one of the input parameters has length ``1``, it is tiled so that its length matches
-    that of the other one. For example, passing ``r=[1,2,3]`` and ``delta=1`` is equivalent to
-    passing ``r=[1,2,3]`` and ``delta=[1,1,1]``.
+    The ``(A, b, c)`` triple of a squeezing gate.
 
     Args:
         r: The squeezing magnitudes.
         delta: The squeezing angles.
 
     Returns:
-        The ``(A, b, c)`` triple of the squeezing gates.
+        The ``(A, b, c)`` triple of the squeezing gate.
     """
-    r, delta = _reshape(r=r, delta=delta)
-    n_modes = len(delta)
+    batch_size, batch_dim = _compute_batch_size(r, delta)
+    batch_shape = batch_size or (1,)
 
-    tanhr = math.diag(math.sinh(r) / math.cosh(r))
-    sechr = math.diag(1 / math.cosh(r))
+    r = math.broadcast_to(r, batch_shape, dtype=math.complex128)
+    delta = math.broadcast_to(delta, batch_shape, dtype=math.complex128)
 
-    A = math.block([[-math.exp(1j * delta) * tanhr, sechr], [sechr, math.exp(-1j * delta) * tanhr]])
-    b = _vacuum_B_vector(n_modes * 2)
-    c = math.prod(1 / math.sqrt(math.cosh(r)))
+    tanhr = math.sinh(r) / math.cosh(r)
+    sechr = 1 / math.cosh(r)
 
-    return A, b, c
+    A = math.stack(
+        [
+            math.stack([-math.exp(1j * delta) * tanhr, sechr], batch_dim),
+            math.stack([sechr, math.exp(-1j * delta) * tanhr], batch_dim),
+        ],
+        batch_dim,
+    )
+    b = math.broadcast_to(_vacuum_B_vector(2), batch_shape + (2,))
+    c = math.cast(1 / math.sqrt(math.cosh(r)), math.complex128)
+
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def beamsplitter_gate_Abc(
-    theta: Union[float, Iterable[float]], phi: Union[float, Iterable[float]] = 0
-) -> Union[Matrix, Vector, Scalar]:
+    theta: float | Iterable[float], phi: float | Iterable[float] = 0
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of two-mode beamsplitter gates.
-
-    The number of modes depends on the length of the input parameters.
-
-    If one of the input parameters has length ``1``, it is tiled so that its length matches
-    that of the other one. For example, passing ``theta=[1,2,3]`` and ``phi=1`` is equivalent to
-    passing ``theta=[1,2,3]`` and ``phi=[1,1,1]``.
+    The ``(A, b, c)`` triple of a tensor product of a two-mode beamsplitter gate.
 
     Args:
         theta: The transmissivity parameters.
         phi: The phase parameters.
 
     Returns:
-        The ``(A, b, c)`` triple of the beamsplitter gates.
+        The ``(A, b, c)`` triple of the beamsplitter gate.
     """
-    theta, phi = _reshape(theta=theta, phi=phi)
-    n_modes = 2 * len(theta)
+    batch_size, batch_dim = _compute_batch_size(theta, phi)
+    batch_shape = batch_size or (1,)
 
-    O_n = math.zeros((n_modes, n_modes), math.complex128)
-    costheta = math.diag(math.cos(theta))
-    sintheta = math.diag(math.sin(theta))
-    V = math.block(
+    theta = math.broadcast_to(theta, batch_shape, dtype=math.complex128)
+    phi = math.broadcast_to(phi, batch_shape, dtype=math.complex128)
+
+    O_matrix = math.zeros(batch_shape + (2, 2), math.complex128)
+    costheta = math.cos(theta)
+    sintheta = math.sin(theta)
+
+    V = math.stack(
         [
-            [costheta, -math.exp(-1j * phi) * sintheta],
-            [math.exp(1j * phi) * sintheta, costheta],
-        ]
+            math.stack([costheta, -math.exp(math.astensor(-1j) * phi) * sintheta], batch_dim),
+            math.stack([math.exp(math.astensor(1j) * phi) * sintheta, costheta], batch_dim),
+        ],
+        batch_dim,
     )
 
-    A = math.block([[O_n, V], [math.transpose(V), O_n]])
-    b = _vacuum_B_vector(n_modes * 2)
-    c = 1.0 + 0j
+    perm = tuple(range(len(V.shape)))
+    perm = perm[:batch_dim] + perm[batch_dim:][::-1]
 
-    return A, b, c
+    A = math.concat(
+        [math.concat([O_matrix, V], -1), math.concat([math.transpose(V, perm), O_matrix], -1)], -2
+    )
+    b = math.broadcast_to(_vacuum_B_vector(4), batch_shape + (4,))
+    c = math.ones(batch_shape, math.complex128)
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def twomode_squeezing_gate_Abc(
-    r: Union[float, Iterable[float]], phi: Union[float, Iterable[float]] = 0
-) -> Union[Matrix, Vector, Scalar]:
+    r: float | Iterable[float], phi: float | Iterable[float] = 0
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of two-mode squeezing gates.
-
-    The number of modes depends on the length of the input parameters.
-
-    If one of the input parameters has length ``1``, it is tiled so that its length matches
-    that of the other one. For example, passing ``r=[1,2,3]`` and ``phi=1`` is equivalent to
-    passing ``r=[1,2,3]`` and ``phi=[1,1,1]``.
+    The ``(A, b, c)`` triple of a tensor product of a two-mode squeezing gate.
 
     Args:
         r: The squeezing magnitudes.
         phi: The squeezing phase.
 
     Returns:
-        The ``(A, b, c)`` triple of the two mode squeezing gates.
+        The ``(A, b, c)`` triple of the two mode squeezing gate.
     """
-    r, phi = _reshape(r=r, phi=phi)
-    n_modes = 2 * len(r)
+    batch_size, batch_dim = _compute_batch_size(r, phi)
+    batch_shape = batch_size or (1,)
 
-    O = math.zeros((len(r), len(r)), math.complex128)
-    tanhr = math.diag(math.exp(1j * phi) * math.sinh(r) / math.cosh(r))
-    sechr = math.diag(1 / math.cosh(r))
+    r = math.broadcast_to(r, batch_shape, dtype=math.complex128)
+    phi = math.broadcast_to(phi, batch_shape, dtype=math.complex128)
 
-    A_block1 = math.block([[O, tanhr], [tanhr, O]])
+    O_matrix = math.zeros(batch_shape, math.complex128)
+    tanhr = math.exp(1j * phi) * math.sinh(r) / math.cosh(r)
+    sechr = 1 / math.cosh(r)
 
-    A_block2 = math.block([[O, -math.conj(tanhr)], [-math.conj(tanhr), O]])
+    A_block1 = math.stack(
+        [math.stack([O_matrix, tanhr], batch_dim), math.stack([tanhr, O_matrix], batch_dim)],
+        batch_dim,
+    )
+    A_block2 = math.stack(
+        [
+            math.stack([O_matrix, -math.conj(tanhr)], batch_dim),
+            math.stack([-math.conj(tanhr), O_matrix], batch_dim),
+        ],
+        batch_dim,
+    )
+    A_block3 = math.stack(
+        [math.stack([sechr, O_matrix], batch_dim), math.stack([O_matrix, sechr], batch_dim)],
+        batch_dim,
+    )
 
-    A_block3 = math.block([[sechr, O], [O, sechr]])
+    A = math.concat(
+        [math.concat([A_block1, A_block3], -1), math.concat([A_block3, A_block2], -1)], -2
+    )
+    b = math.broadcast_to(_vacuum_B_vector(4), batch_shape + (4,))
+    c = math.cast(1 / math.cosh(r), math.complex128)
 
-    A = math.block([[A_block1, A_block3], [A_block3, A_block2]])
-    b = _vacuum_B_vector(n_modes * 2)
-    c = math.prod(1 / math.cosh(r))
-
-    return A, b, c
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
-def identity_Abc(n_modes: int) -> Union[Matrix, Vector, Scalar]:
+def identity_Abc(n_modes: int) -> tuple[Matrix, Vector, Scalar]:
     r"""
     The ``(A, b, c)`` triple of a tensor product of identity gates.
 
@@ -563,92 +607,91 @@ def identity_Abc(n_modes: int) -> Union[Matrix, Vector, Scalar]:
 # ~~~~~~~~~~
 
 
-def attenuator_Abc(eta: Union[float, Iterable[float]]) -> Union[Matrix, Vector, Scalar]:
+def attenuator_Abc(eta: float | Iterable[float]) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of of a tensor product of atternuators.
-
-    The number of modes depends on the length of the input parameters.
+    The ``(A, b, c)`` triple of an attenuator.
 
     Args:
         eta: The values of the transmissivities.
 
     Returns:
-        The ``(A, b, c)`` triple of the attenuator channels.
+        The ``(A, b, c)`` triple of the attenuator channel.
 
     Raises:
         ValueError: If ``eta`` is larger than `1` or smaller than `0`.
     """
-    eta = math.atleast_1d(eta, math.complex128)
-    n_modes = len(eta)
+    batch_size, batch_dim = _compute_batch_size(eta)
+    batch_shape = batch_size or (1,)
 
-    math.error_if(eta, math.real(eta) > 1, f"Transmissivity is {eta}, must be <= 1")
-    math.error_if(eta, math.real(eta) < 0, f"Transmissivity is {eta}, must be >= 0")
+    eta = math.broadcast_to(eta, batch_shape, dtype=math.complex128)
 
-    O_n = math.zeros((n_modes, n_modes), math.complex128)
-    eta1 = math.reshape(math.diag(math.sqrt(eta)), (n_modes, n_modes))
-    eta2 = math.eye(n_modes, math.complex128) - math.reshape(math.diag(eta), (n_modes, n_modes))
+    math.error_if(eta, math.any(math.real(eta) > 1), "Found transmissivity greater than `1`.")
+    math.error_if(eta, math.any(math.real(eta) < 0), "Found transmissivity less than `0`.")
 
-    A = math.block(
+    O_matrix = math.zeros(batch_shape, math.complex128)
+    eta1 = math.sqrt(eta)
+    eta2 = 1 - eta
+
+    A = math.stack(
         [
-            [O_n, eta1, O_n, O_n],
-            [eta1, O_n, O_n, eta2],
-            [O_n, O_n, O_n, eta1],
-            [O_n, eta2, eta1, O_n],
-        ]
+            math.stack([O_matrix, eta1, O_matrix, O_matrix], batch_dim),
+            math.stack([eta1, O_matrix, O_matrix, eta2], batch_dim),
+            math.stack([O_matrix, O_matrix, O_matrix, eta1], batch_dim),
+            math.stack([O_matrix, eta2, eta1, O_matrix], batch_dim),
+        ],
+        batch_dim,
     )
+    b = math.broadcast_to(_vacuum_B_vector(4), batch_shape + (4,))
+    c = math.ones(batch_shape, math.complex128)
 
-    b = _vacuum_B_vector(n_modes * 4)
-    c = 1.0 + 0j
-
-    return A, b, c
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
-def amplifier_Abc(g: Union[float, Iterable[float]]) -> Union[Matrix, Vector, Scalar]:
+def amplifier_Abc(g: float | Iterable[float]) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of amplifiers.
-
-    The number of modes depends on the length of the input parameters.
+    The ``(A, b, c)`` triple of an amplifier.
 
     Args:
         g: The values of the gains.
 
     Returns:
-        The ``(A, b, c)`` triple of the amplifier channels.
+        The ``(A, b, c)`` triple of the amplifier channel.
 
     Raises:
         ValueError: If ``g`` is smaller than `1`.
     """
-    g = math.atleast_1d(g, math.complex128)
-    n_modes = len(g)
+    batch_size, batch_dim = _compute_batch_size(g)
+    batch_shape = batch_size or (1,)
 
-    for g_val in g:
-        math.error_if(
-            g_val, math.real(g_val) < 1, "Found amplifier with gain ``g`` smaller than `1`."
-        )
+    g = math.broadcast_to(g, batch_shape, dtype=math.complex128)
 
-    O_n = math.zeros((n_modes, n_modes), math.complex128)
-    g1 = math.reshape(math.diag(1 / math.sqrt(g)), (n_modes, n_modes))
-    g2 = math.reshape(math.diag(1 - 1 / g), (n_modes, n_modes))
-    A = math.block(
-        [
-            [O_n, g1, g2, O_n],
-            [g1, O_n, O_n, O_n],
-            [g2, O_n, O_n, g1],
-            [O_n, O_n, g1, O_n],
-        ]
+    math.error_if(
+        g, math.any(math.real(g) < 1), "Found amplifier with gain ``g`` smaller than `1`."
     )
 
-    b = _vacuum_B_vector(n_modes * 4)
-    c = math.prod(1 / g)
+    O_matrix = math.zeros(batch_shape, math.complex128)
+    g1 = 1 / math.sqrt(g)
+    g2 = 1 - 1 / g
+    A = math.stack(
+        [
+            math.stack([O_matrix, g1, g2, O_matrix], batch_dim),
+            math.stack([g1, O_matrix, O_matrix, O_matrix], batch_dim),
+            math.stack([g2, O_matrix, O_matrix, g1], batch_dim),
+            math.stack([O_matrix, O_matrix, g1, O_matrix], batch_dim),
+        ],
+        batch_dim,
+    )
+    b = math.broadcast_to(_vacuum_B_vector(4), batch_shape + (4,))
+    c = math.cast(1 / g, math.complex128)
 
-    return A, b, c
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def fock_damping_Abc(
-    beta: Union[float, Iterable[float]],
-) -> Union[Matrix, Vector, Scalar]:
+    beta: float | Iterable[float],
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
-    The ``(A, b, c)`` triple of a tensor product of Fock dampers.
+    The ``(A, b, c)`` triple of a Fock damper.
 
     Args:
         beta: The damping parameter.
@@ -656,20 +699,24 @@ def fock_damping_Abc(
     Returns:
         The ``(A, b, c)`` triple of the Fock damping operator.
     """
-    beta = math.atleast_1d(beta, math.complex128)
-    n_modes = len(beta)
+    batch_size, batch_dim = _compute_batch_size(beta)
+    batch_shape = batch_size or (1,)
 
-    O_n = math.zeros((n_modes, n_modes), math.complex128)
-    B_n = math.diag(math.astensor(math.exp(-beta))).reshape((n_modes, n_modes))
+    beta = math.broadcast_to(beta, batch_shape, dtype=math.complex128)
 
-    A = math.block([[O_n, B_n], [B_n, O_n]])
-    b = _vacuum_B_vector(n_modes * 2)
-    c = 1.0 + 0j
+    O_matrix = math.zeros(batch_shape, math.complex128)
+    B_n = math.exp(-beta)
 
-    return A, b, c
+    A = math.stack(
+        [math.stack([O_matrix, B_n], batch_dim), math.stack([B_n, O_matrix], batch_dim)], batch_dim
+    )
+    b = math.broadcast_to(_vacuum_B_vector(2), batch_shape + (2,))
+    c = math.ones(batch_shape, math.complex128)
+
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
-def gaussian_random_noise_Abc(Y: RealMatrix) -> Union[Matrix, Vector, Scalar]:
+def gaussian_random_noise_Abc(Y: RealMatrix) -> tuple[Matrix, Vector, Scalar]:
     r"""
     The triple (A, b, c) for the gaussian random noise channel.
 
@@ -679,6 +726,11 @@ def gaussian_random_noise_Abc(Y: RealMatrix) -> Union[Matrix, Vector, Scalar]:
     Returns:
         The ``(A, b, c)`` triple of the Gaussian random noise channel.
     """
+    batch_size = Y.shape[:-2]
+    batch_shape = batch_size or (1,)
+
+    Y = math.broadcast_to(Y, batch_shape + Y.shape[-2:], dtype=math.complex128)
+
     m = Y.shape[-1] // 2
     xi = math.eye(2 * m, dtype=math.complex128) + Y / settings.HBAR
     xi_inv = math.inv(xi)
@@ -715,13 +767,15 @@ def gaussian_random_noise_Abc(Y: RealMatrix) -> Union[Matrix, Vector, Scalar]:
     )
 
     A = math.Xmat(2 * m) @ R @ xi_inv_in_blocks @ math.conj(R).T
-    b = math.zeros(4 * m)
+    b = math.zeros(batch_shape + (4 * m,))
     c = 1 / math.sqrt(math.det(xi))
 
-    return A, b, c
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
-def bargmann_to_quadrature_Abc(n_modes: int, phi: float) -> tuple[Matrix, Vector, Scalar]:
+def bargmann_to_quadrature_Abc(
+    n_modes: int, phi: float | Iterable[float]
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
     The ``(A, b, c)`` triple of the multi-mode kernel :math:`\langle \vec{p}|\vec{z} \rangle` between bargmann representation with ABC Ansatz form and quadrature representation with ABC Ansatz.
     The kernel can be considered as a Unitary-like component: the out_ket wires are related to the real variable :math:`\vec{p}` in quadrature representation and the in_ket wires are related to the complex variable :math:`\vec{z}`.
@@ -735,19 +789,30 @@ def bargmann_to_quadrature_Abc(n_modes: int, phi: float) -> tuple[Matrix, Vector
     Returns:
         The ``(A, b, c)`` triple of the map from bargmann representation with ABC Ansatz to quadrature representation with ABC Ansatz.
     """
+    batch_size, batch_dim = _compute_batch_size(phi)
+    batch_shape = batch_size or (1,)
+
+    phi = math.broadcast_to(phi, batch_shape, dtype=math.complex128)
+
     hbar = settings.HBAR
-    Id = np.eye(n_modes, dtype=np.complex128)
-    e = np.exp(-1j * phi + 1j * np.pi / 2)
-    A = np.kron(
-        [
-            [-1 / hbar, -1j * e * np.sqrt(2 / hbar)],
-            [-1j * e * np.sqrt(2 / hbar), e * e],
-        ],
+    Id = math.eye(n_modes, dtype=math.complex128)
+    e = math.exp(-1j * phi + 1j * np.pi / 2)
+    A = math.kron(
+        math.stack(
+            [
+                math.stack(
+                    [math.broadcast_to(-1 / hbar, batch_shape), -1j * e * np.sqrt(2 / hbar)],
+                    batch_dim,
+                ),
+                math.stack([-1j * e * np.sqrt(2 / hbar), e * e], batch_dim),
+            ],
+            batch_dim,
+        ),
         Id,
     )
-    b = _vacuum_B_vector(2 * n_modes)
-    c = (1.0 + 0j) / (np.pi * hbar) ** (0.25 * n_modes)
-    return A, b, c
+    b = math.broadcast_to(_vacuum_B_vector(2 * n_modes), batch_shape + (2 * n_modes,))
+    c = math.ones(batch_shape, math.complex128) / (np.pi * hbar) ** (0.25 * n_modes)
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -755,7 +820,7 @@ def bargmann_to_quadrature_Abc(n_modes: int, phi: float) -> tuple[Matrix, Vector
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 
-def displacement_map_s_parametrized_Abc(s: int, n_modes: int) -> Union[Matrix, Vector, Scalar]:
+def displacement_map_s_parametrized_Abc(s: int, n_modes: int) -> tuple[Matrix, Vector, Scalar]:
     r"""
     The ``(A, b, c)`` triple of a multi-mode ``s``\-parametrized displacement map.
     :math:
@@ -769,11 +834,35 @@ def displacement_map_s_parametrized_Abc(s: int, n_modes: int) -> Union[Matrix, V
     Returns:
         The ``(A, b, c)`` triple of the multi-mode ``s``-parametrized dispalcement map :math:`D_s(\gamma)`.
     """
-    A = math.block(
+    batch_size, _ = _compute_batch_size(s)
+    batch_shape = batch_size or (1,)
+
+    s = math.broadcast_to(s, batch_shape, dtype=math.complex128)
+
+    A = math.concat(
         [
-            [(s - 1) / 2 * math.Xmat(num_modes=n_modes), -math.Zmat(num_modes=n_modes)],
-            [-math.Zmat(num_modes=n_modes), math.Xmat(num_modes=n_modes)],
-        ]
+            np.concat(
+                [
+                    (s[..., None, None] - 1) / 2 * math.Xmat(num_modes=n_modes),
+                    math.broadcast_to(
+                        -math.Zmat(num_modes=n_modes), batch_shape + (2 * n_modes, 2 * n_modes)
+                    ),
+                ],
+                -1,
+            ),
+            np.concat(
+                [
+                    math.broadcast_to(
+                        -math.Zmat(num_modes=n_modes), batch_shape + (2 * n_modes, 2 * n_modes)
+                    ),
+                    math.broadcast_to(
+                        math.Xmat(num_modes=n_modes), batch_shape + (2 * n_modes, 2 * n_modes)
+                    ),
+                ],
+                -1,
+            ),
+        ],
+        -2,
     )
     order_list = math.arange(4 * n_modes)  # [0,3,1,2]
     order_list = list(
@@ -789,10 +878,10 @@ def displacement_map_s_parametrized_Abc(s: int, n_modes: int) -> Union[Matrix, V
         )
     )
 
-    A = math.astensor(math.asnumpy(A)[order_list, :][:, order_list])
-    b = _vacuum_B_vector(4 * n_modes)
-    c = 1.0
-    return math.astensor(A), b, c
+    A = math.astensor(math.asnumpy(A)[..., order_list, :][..., :, order_list])
+    b = math.broadcast_to(_vacuum_B_vector(4 * n_modes), batch_shape + (4 * n_modes,))
+    c = math.ones(batch_shape, math.complex128)
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
 def complex_fourier_transform_Abc(n_modes: int) -> tuple[Matrix, Vector, Scalar]:
@@ -822,7 +911,7 @@ def complex_fourier_transform_Abc(n_modes: int) -> tuple[Matrix, Vector, Scalar]
 # ~~~~~~~~~~~~~~~~
 
 
-def attenuator_kraus_Abc(eta: float) -> Union[Matrix, Vector, Scalar]:
+def attenuator_kraus_Abc(eta: float | Iterable[float]) -> tuple[Matrix, Vector, Scalar]:
     r"""
     The entire family of Kraus operators of the attenuator (loss) channel as a single ``(A, b, c)`` triple.
     The last index is the "bond" index which should be summed/integrated over.
@@ -833,18 +922,32 @@ def attenuator_kraus_Abc(eta: float) -> Union[Matrix, Vector, Scalar]:
     Returns:
         The ``(A, b, c)`` triple of the kraus operators of the attenuator (loss) channel.
     """
+    batch_size, batch_dim = _compute_batch_size(eta)
+    batch_shape = batch_size or (1,)
+
+    eta = math.broadcast_to(eta, batch_shape, dtype=math.complex128)
+
     costheta = math.sqrt(eta)
     sintheta = math.sqrt(1 - eta)
 
-    A = math.astensor(
-        [[0, costheta, 0], [costheta, 0, -sintheta], [0, -sintheta, 0]], math.complex128
+    O_matrix = math.zeros(batch_shape, math.complex128)
+
+    A = math.stack(
+        [
+            math.stack([O_matrix, costheta, O_matrix], batch_dim),
+            math.stack([costheta, O_matrix, -sintheta], batch_dim),
+            math.stack([O_matrix, -sintheta, O_matrix], batch_dim),
+        ],
+        batch_dim,
     )
-    b = _vacuum_B_vector(3)
-    c = 1.0 + 0j
-    return A, b, c
+    b = math.broadcast_to(_vacuum_B_vector(3), batch_shape + (3,))
+    c = math.ones(batch_shape, math.complex128)
+    return A if batch_size else A[0], b if batch_size else b[0], c if batch_size else c[0]
 
 
-def XY_to_channel_Abc(X: RealMatrix, Y: RealMatrix, d: Vector | None = None) -> ComplexMatrix:
+def XY_to_channel_Abc(
+    X: RealMatrix, Y: RealMatrix, d: Vector | None = None
+) -> tuple[Matrix, Vector, Scalar]:
     r"""
     The method to compute the A matrix of a channel based on its X, Y, and d.
     Args:
