+++ tests/test_lab_dev/test_circuit_components_utils/test_b_to_ps.py
@@ -23,7 +23,7 @@ from mrmustard import math
 from mrmustard.lab_dev import DM, BtoPS, Identity, Ket
 from mrmustard.physics.bargmann_utils import wigner_to_bargmann_rho
 from mrmustard.physics.gaussian_integrals import complex_gaussian_integral_2
-from mrmustard.physics.triples_batched import displacement_map_s_parametrized_Abc
+from mrmustard.physics.triples import displacement_map_s_parametrized_Abc
 
 
 class TestBtoPS:
