+++ tests/test_lab_dev/test_circuit_components_utils/test_trace_out.py
@@ -24,7 +24,7 @@ from mrmustard.lab_dev.circuit_components import CircuitComponent
 from mrmustard.lab_dev.circuit_components_utils import TraceOut
 from mrmustard.lab_dev.states import Coherent
 from mrmustard.physics.ansatz import PolyExpAnsatz
-from mrmustard.physics.triples_batched import identity_Abc
+from mrmustard.physics.triples import identity_Abc
 from mrmustard.physics.wires import Wires
